requirements.txt
numpy
pandas
matplotlib
tensorflow
tensorflow-federated-nightly